[Chore] Separate out `vllm.utils.hardware_utils.py`
