+++ tests/kernels/core/test_uva.py
@@ -3,7 +3,7 @@
 import pytest
 import torch
 
-from vllm.utils.hardware_utils import is_uva_available
+from vllm.utils.platform_utils import is_uva_available
 from vllm.utils.torch_utils import get_cuda_view_from_cpu_tensor
 
 CUDA_DEVICES = [f"cuda:{i}" for i in range(1 if torch.cuda.device_count() == 1 else 2)]
