+++ vllm/lora/lora_weights.py
@@ -8,7 +8,7 @@ import torch
 import torch.types
 
 from vllm.lora.peft_helper import PEFTHelper
-from vllm.utils.hardware_utils import is_pin_memory_available
+from vllm.utils.platform_utils import is_pin_memory_available
 
 
 class LoRALayerWeights:
