+++ vllm/v1/worker/gpu_model_runner.py
@@ -75,10 +75,10 @@ from vllm.utils import (
     length_from_prompt_token_ids_or_embeds,
     round_up,
 )
-from vllm.utils.hardware_utils import is_pin_memory_available
 from vllm.utils.jsontree import json_map_leaves
 from vllm.utils.mem_constants import GiB_bytes
 from vllm.utils.mem_utils import DeviceMemoryProfiler
+from vllm.utils.platform_utils import is_pin_memory_available
 from vllm.utils.torch_utils import (
     get_dtype_size,
     kv_cache_dtype_str_to_dtype,
